# src/shippingservice/genproto/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "demo.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/currencies": {
      "get": {
        "operationId": "CurrencyService_GetSupportedCurrencies",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/hipstershopGetSupportedCurrenciesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "CurrencyService"
        ]
      }
    },
    "/api/currencies/convert": {
      "post": {
        "operationId": "CurrencyService_Convert",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/hipstershopMoney"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/hipstershopCurrencyConversionRequest"
            }
          }
        ],
        "tags": [
          "CurrencyService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "hipstershopCurrencyConversionRequest": {
      "type": "object",
      "properties": {
        "from": {
          "$ref": "#/definitions/hipstershopMoney"
        },
        "toCode": {
          "type": "string",
          "description": "The 3-letter currency code defined in ISO 4217."
        }
      }
    },
...
    "hipstershopGetSupportedCurrenciesResponse": {
      "type": "object",
      "properties": {
        "currencyCodes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "The 3-letter currency code defined in ISO 4217."
        }
      }
    },
...
    "hipstershopMoney": {
      "type": "object",
      "properties": {
        "currencyCode": {
          "type": "string",
          "description": "The 3-letter currency code defined in ISO 4217."
        },
        "units": {
          "type": "string",
          "format": "int64",
          "description": "The whole units of the amount.\nFor example if `currencyCode` is `\"USD\"`, then 1 unit is one US dollar."
        },
        "nanos": {
          "type": "integer",
          "format": "int32",
          "description": "Number of nano (10^-9) units of the amount.\nThe value must be between -999,999,999 and +999,999,999 inclusive.\nIf `units` is positive, `nanos` must be positive or zero.\nIf `units` is zero, `nanos` can be positive, zero, or negative.\nIf `units` is negative, `nanos` must be negative or zero.\nFor example $-1.75 is represented as `units`=-1 and `nanos`=-750,000,000."
        }
      },
      "description": "Represents an amount of money with its currency type."
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}